some minor changes (#56)
