+++ x-pack/plugins/cloud_security_posture/public/pages/compliance_dashboard/compliance_charts/compliance_trend_chart.tsx
@@ -14,14 +14,13 @@ import {
   niceTimeFormatByDay,
   AreaSeries,
 } from '@elastic/charts';
-import { dateValueToTuple } from '../index';
-// TODO: find out how to import from the server folder without warnings
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { BenchmarkStats } from '../../../../../server/cloud_posture/types';
+import type { BenchmarkStats } from '../../../../common/types';
 
 interface ComplianceTrendChartProps {
   data: BenchmarkStats;
 }
+
+export const dateValueToTuple = ({ date, value }: { date: number; value: number }) => [date, value];
 
 export const ComplianceTrendChart = ({ data: { postureScore } }: ComplianceTrendChartProps) => {
   if (postureScore === undefined) return null;
