+++ x-pack/plugins/cloud_security_posture/public/pages/compliance_dashboard/compliance_dashboard.tsx
@@ -1,0 +1,45 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+import { EuiSpacer, EuiTitle } from '@elastic/eui';
+import { SummarySection } from './dashboard_sections/summary_section';
+import { BenchmarksSection } from './dashboard_sections/benchmarks_section';
+import { useCloudPostureStatsApi } from '../../common/api';
+import { CspPageTemplate } from '../../components/page_template';
+
+const CompliancePage = () => {
+  const getStats = useCloudPostureStatsApi();
+  if (getStats.isLoading) return null;
+
+  return (
+    <>
+      <EuiTitle>
+        <h3>{'Summary'}</h3>
+      </EuiTitle>
+      <EuiSpacer />
+      <SummarySection />
+      <EuiSpacer />
+      <EuiTitle>
+        <h3>{'Benchmarks'}</h3>
+      </EuiTitle>
+      <EuiSpacer />
+      <BenchmarksSection />
+      <EuiSpacer />
+    </>
+  );
+};
+
+export const ComplianceDashboard = () => (
+  <CspPageTemplate
+    pageHeader={{
+      pageTitle: 'Compliance',
+    }}
+  >
+    <CompliancePage />
+  </CspPageTemplate>
+);
