+++ x-pack/plugins/cloud_security_posture/public/pages/compliance_dashboard/index.tsx
@@ -5,43 +5,4 @@
  * 2.0.
  */
 
-import React from 'react';
-import { EuiSpacer, EuiTitle } from '@elastic/eui';
-import { SummarySection } from './dashboard_sections/summary_section';
-import { BenchmarksSection } from './dashboard_sections/benchmarks_section';
-import { useCloudPostureStatsApi } from '../../common/api';
-import { CspPageTemplate } from '../../components/page_template';
-
-export const dateValueToTuple = ({ date, value }: { date: number; value: number }) => [date, value];
-
-const CompliancePage = () => {
-  const getStats = useCloudPostureStatsApi();
-  if (getStats.isLoading) return null;
-
-  return (
-    <>
-      <EuiTitle>
-        <h3>{'Summary'}</h3>
-      </EuiTitle>
-      <EuiSpacer />
-      <SummarySection />
-      <EuiSpacer />
-      <EuiTitle>
-        <h3>{'Benchmarks'}</h3>
-      </EuiTitle>
-      <EuiSpacer />
-      <BenchmarksSection />
-      <EuiSpacer />
-    </>
-  );
-};
-
-export const ComplianceDashboard = () => (
-  <CspPageTemplate
-    pageHeader={{
-      pageTitle: 'Compliance',
-    }}
-  >
-    <CompliancePage />
-  </CspPageTemplate>
-);
+export * from './compliance_dashboard';
